XDMA: libxdma.c: modify user interrupt handling for UARTLite over PCIe
Ignoring flag checking of the user irqs and
enabling UARTLite irq handling in any conditions.

The edge-sensitive interrupt signals generated by the UARTLite
controller is inconsistent. As a result, the interrupt signal would be
dismissed without clearing the pending state of the UARTLite register,
leading to undetected user irq signals in the XDMA MSI interrupt
handler.

With such an optimization, the interrupt handler of the UARTLite driver
would be triggered by the XDMA MSI interrupt handler in any conditions,
without checking the status of user irq signal lines in the XDMA engine.