xdma: uart: fix user irq handler
We assume that the UART controller occupies the user irq line #0
of the XDMA instance. Before the user irq handler registered via
xdma_usr_isr_register() in the xdma_uart driver,
it is ensured that no user_irq_service() handler would be called
in the MSI irq handler of XDMA engine.